+++ src/AppDesktop.js
@@ -12,6 +12,7 @@ import { CMap } from './views/map';
 import { NavbarChangeset } from './views/navbar_changeset';
 import { NavbarSidebar } from './views/navbar_sidebar';
 import { Home } from './views/home';
+import { Authorized } from './views/authorized';
 import { Modal } from './views/modal';
 import { User } from './views/user';
 import { SavedFilters } from './views/saved_filters';
@@ -61,11 +62,12 @@ export const AppDesktop = () => {
                 <Route path={`${BASE_PATH}/saved-filters`} component={SavedFilters} />
                 <Route path={`${BASE_PATH}/trusted-users`} component={TrustedUsers} />
                 <Route path={`${BASE_PATH}/watchlist`} component={Watchlist} />
-              </CSSTransitionGroup>
+                <Route path={`${BASE_PATH}/authorized`} component={Authorized} />
+              </CSSTransitionGroup >
             )}
           />
-        </div>
-      </div>
+        </div >
+      </div >
       <Modal />
     </>
   );