+++ src/app.js
@@ -21,6 +21,13 @@ export class App extends Component {
           .keySeq()
           .sort((a, b) => a.localeCompare(b))
           .join(',');
+        gaPageView({
+          page: `/?filters=${filters}`,
+          title: 'Filters',
+          hitType: 'pageview'
+        });
+      } else {
+        gaPageView({ page: '/', title: 'Home', hitType: 'pageview' });
       }
     }
   }