+++ src/index.js
@@ -3,6 +3,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { ConnectedRouter } from 'react-router-redux';
 import { Provider } from 'react-redux';
+import ReactGA from 'react-ga4';
+import Raven from 'raven-js';
 import { history } from './store/history';
 import { store } from './store';
 import { unregisterServiceWorker } from './serviceworker';
@@ -14,6 +16,24 @@ import 'react-select/dist/react-select.css';
 
 import { App } from './app';
 
+ReactGA.initialize([
+  {
+    trackingId: 'G-05Y4GLDM03',
+    gaOptions: {
+      anonymizeIp: true,
+      screenResolution: `${window.screen.availWidth}X${window.screen.availHeight}`,
+      appName: stack
+    }
+  }
+]);
+
+if (process.env.NODE_ENV === 'production') {
+  Raven.config('https://5637ef87f5794e2fb9e1e5fe9119688d@sentry.io/175926', {
+    release: appVersion,
+    environment: process.env.NODE_ENV,
+    debug: isDev
+  }).install();
+}
 // if (process.env.NODE_ENV !== 'production') {
 // const { whyDidYouUpdate } = require('why-did-you-update');
 // whyDidYouUpdate(React);