+++ src/network/mapping_team.js
@@ -15,9 +15,7 @@ export function createMappingTeam(token: string, name: string, users: string) {
     })
   })
     .then(handleErrors)
-    .then(res => {
-      return res.json();
-    });
+    .then(res => res.json());
 }
 
 export function fetchMappingTeam(token: string, id: number) {
@@ -30,9 +28,7 @@ export function fetchMappingTeam(token: string, id: number) {
     }
   })
     .then(handleErrors)
-    .then(res => {
-      return res.json();
-    });
+    .then(res => res.json());
 }
 export function deleteMappingTeam(token: string, id: number) {
   return fetch(`${API_URL}/mapping-team/${id}/`, {
@@ -55,9 +51,8 @@ export function fetchUserMappingTeams(token: string, id: number) {
     }
   })
     .then(handleErrors)
-    .then(res => {
-      return res.json();
-    });
+    .then(res => res.json())
+    .then(res => res.results);
 }
 
 export function fetchTrustedMappingTeams(token: string) {
@@ -70,9 +65,7 @@ export function fetchTrustedMappingTeams(token: string) {
     }
   })
     .then(handleErrors)
-    .then(res => {
-      return res.json();
-    });
+    .then(res => res.json());
 }
 
 export function updateMappingTeam(
@@ -94,7 +87,5 @@ export function updateMappingTeam(
     })
   })
     .then(handleErrors)
-    .then(res => {
-      return res.json();
-    });
+    .then(res => res.json());
 }